Did more stuff

+++ processTemps.h
@@ -6,7 +6,7 @@
 #include "parseTemps.h"
 
 std::map<int, std::vector<float>> convertToMap(std::vector<CoreTempReading> readings);
-float calcSlope(std::map<int, std::vector<float>> m, int key);
-float calcIntercept(std::map<int, std::vector<float>> m, float slope, int key);
+double calcSlope(std::map<int, std::vector<float>> m, int key);
+double calcIntercept(std::map<int, std::vector<float>> m, double slope, int key);
 
 #endif // PROCESSTEMPS_H_INCLUDED
